refactor: Renamed variables, classes for readability, restructed libraries
