+++ ArduinoSerialParser/Libraries/Averager.h
